requirements.txt
requests
bs4
ndp_ep
pandas
msgpack
blosc
kafka-python
confluent-kafka
python-dotenv
scidx_streaming
matplotlib
seaborn>=0.13
lz4
plotly
nbformat
microsoft-aurora
cdsapi
fastmcp
folium
sage_data_client
herbie-data
numpy
scipy
websocket-client
cartopy
shapely
pyproj
fastapi
uvicorn[standard]
pydantic
redis>=5.0
rq>=1.16
